+++ internal/configuration/validation_test.go
